Restructured Notes and Contacts, added utils and moved constants to closest scope
