+++ app/javascript/src/components/Dashboard/Contacts/Pane/Create.jsx
@@ -3,7 +3,6 @@ import React from "react";
 import { Pane, Typography } from "neetoui";
 import { useTranslation } from "react-i18next";
 
-import { CONTACTS_FORM_INITIAL_VALUES } from "./constants";
 import Form from "./Form";
 
 const Create = ({ showPane, setShowPane }) => {
@@ -18,11 +17,7 @@ const Create = ({ showPane, setShowPane }) => {
           {t("create", { entity: "Contact" })}
         </Typography>
       </Pane.Header>
-      <Form
-        contacts={CONTACTS_FORM_INITIAL_VALUES}
-        setShowPane={setShowPane}
-        onClose={onClose}
-      />
+      <Form setShowPane={setShowPane} onClose={onClose} />
     </Pane>
   );
 };
