+++ app/javascript/src/components/Dashboard/Contacts/index.jsx
@@ -1,13 +1,13 @@
 import React, { useState } from "react";
 
-import { MenuVertical } from "neetoicons";
-import { Avatar, Button, Table, Dropdown } from "neetoui";
+import { Button, Table } from "neetoui";
 import { Container, Header } from "neetoui/layouts";
 import { useTranslation } from "react-i18next";
 
 import { CONTACTS_LIST, COLUMN_DATA } from "./constants";
 import DeleteAlert from "./DeleteAlert";
 import NewContactPane from "./Pane/Create";
+import transformContactData from "./utils";
 
 const Contacts = () => {
   const [currentPageNumber, setCurrentPageNumber] = useState(1);
@@ -24,7 +24,7 @@ const Contacts = () => {
         actionBlock={
           <Button
             icon="ri-add-line"
-            label={t("add", { entity: "Contact" })}
+            label={t("header.add", { entity: "Contact" })}
             size="small"
             onClick={() => setShowNewContactPain(true)}
           />
@@ -40,42 +40,7 @@ const Contacts = () => {
         currentPageNumber={currentPageNumber}
         defaultPageSize={10}
         handlePageChange={page => setCurrentPageNumber(page)}
-        rowData={CONTACTS_LIST.map(contact => ({
-          ...contact,
-          icon_button: (
-            <Dropdown
-              className="inline-block"
-              customTarget={<MenuVertical className="mx-auto cursor-pointer" />}
-              onClose={function noRefCheck() {}}
-            >
-              <Dropdown.Menu>
-                <Dropdown.MenuItem.Button>{t("edit")}</Dropdown.MenuItem.Button>
-                <Dropdown.MenuItem.Button
-                  onClick={() => setShowDeleteAlert(true)}
-                >
-                  {t("delete")}
-                </Dropdown.MenuItem.Button>
-              </Dropdown.Menu>
-            </Dropdown>
-          ),
-          name: (
-            <div className="flex items-center" key={contact.id}>
-              <Avatar
-                className="mr-2 inline"
-                size="medium"
-                user={{
-                  imageUrl: "https://i.pravatar.cc/300",
-                  name: "neeto",
-                }}
-                onClick={() => {}}
-              />
-              <div>
-                <span className="block">{contact.name}</span>
-                <span className="block text-gray-500">{contact.role}</span>
-              </div>
-            </div>
-          ),
-        }))}
+        rowData={transformContactData(CONTACTS_LIST, t, setShowDeleteAlert)}
         onRowSelect={() => {}}
       />
       <NewContactPane
