+++ app/javascript/src/components/Dashboard/Notes/Card/utils.js
