+++ app/javascript/src/components/Dashboard/Notes/Pane/Create.jsx
@@ -5,8 +5,6 @@ import { useTranslation } from "react-i18next";
 
 import Form from "./Form";
 
-import { NOTES_FORM_INITIAL_VALUES } from "../constants";
-
 const Create = ({ showPane, setShowPane, notes, setNotes }) => {
   const onClose = () => setShowPane(false);
 
@@ -21,7 +19,6 @@ const Create = ({ showPane, setShowPane, notes, setNotes }) => {
       </Pane.Header>
       <Form
         isEdit={false}
-        note={NOTES_FORM_INITIAL_VALUES}
         notes={notes}
         setNotes={setNotes}
         setShowPane={setShowPane}
